fix: let's move those into the right directory this time
