+++ app/Services/Plugins/PluginInformationService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\Themes;
+namespace App\Services\Plugins;
 
 use App\Models\WpOrg\Plugin;
 use Illuminate\Database\Eloquent\Builder;
