+++ app/Services/Plugins/PluginUpdateService.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\Themes;
+namespace App\Services\Plugins;
 
 use App\Models\WpOrg\Plugin;
 use Illuminate\Support\Collection;
